refactor: 테스트 클래스 패키지의 구조를 프로덕션 패키지와 동일하게 수정
