+++ src/test/java/roomescape/reservation/controller/MemberReservationControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.roomescape.admin;
+package roomescape.reservation.controller;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.hamcrest.Matchers.is;
